+++ feature/intake/src/main/java/com/ashwinrao/packup/intake/viewmodel/FakeIntakeScreenViewModel.kt
@@ -11,7 +11,7 @@ import kotlinx.coroutines.flow.StateFlow
 class FakeIntakeScreenViewModel : IntakeScreenViewModel {
     override val currentItem: StateFlow<Item?> = MutableStateFlow(null)
 
-    override fun fetchCurrentItem(id: Int) {}
+    override fun fetchCurrentItem(id: Long) {}
 
     override fun discardCurrentItem() {}
