+++ feature/intake/src/main/java/com/ashwinrao/packup/intake/viewmodel/IntakeScreenViewModel.kt
@@ -10,7 +10,7 @@ import kotlinx.coroutines.flow.StateFlow
 interface IntakeScreenViewModel {
     val currentItem: StateFlow<Item?>
 
-    fun fetchCurrentItem(id: Int)
+    fun fetchCurrentItem(id: Long)
 
     fun discardCurrentItem()
