+++ feature/main/src/main/java/com/ashwinrao/packup/feature/main/ui/viewmodel/FakeMainScreenViewModel.kt
@@ -13,5 +13,5 @@ class FakeMainScreenViewModel : MainScreenViewModel {
 
     override fun fetchItems() {}
 
-    override fun fetchItem(id: Int) {}
+    override fun fetchItem(id: Long) {}
 }
